1.reformatted cmake files to be more readable
2.updating cmake to allow lto

3.updated git ignore so we can use build and scan-build dirs

4.updated cmake strict flags and added stack protection. updated files to conform to new flags

+++ src/cparser/main.cpp
@@ -972,7 +972,7 @@ s32 main(s32 argc,s8** argv){
     
     auto diff = GetTimeDifferenceMS(start,end);
     
-    printf("TARGET %s PARSE TIME: %f(s)\n",meta_src,diff/1000.0f);
+    printf("TARGET %s PARSE TIME: %f(s)\n",meta_src,(f64)(diff/1000.0f));
     
 #endif
